WaE: -Wsign-compare

+++ src/Observations.h
@@ -69,7 +69,7 @@
         void   get_postseismiclog(std::vector<LogEntry>& postseismiclog_);
         void   get_postseismicexp(std::vector<ExpEntry>& postseismicexp_);
         void   add_offset(double MJD);
-        void   change_offset(int column, double MJD);
+        void   change_offset(size_t column, double MJD);
     };
 
   #endif
